Re-arranging and cleaning code
