+++ image_preprocessing/config/params.yaml
@@ -2,7 +2,7 @@
   ros__parameters:
     rotation_angle: 60
 #   resizeFlag: true
-    grayscaleFlag: false
+    grayscaleFlag: 0
     kernal_size: 0
     amount: 0
     value: 128.0
